alteração no tipo da coluna desafio em monstros

+++ database/migrations/2022_08_24_130027_create_monstros_table.php
@@ -16,7 +16,7 @@ class CreateMonstrosTable extends Migration
         Schema::create('monstros', function (Blueprint $table) {
             $table->id();
             $table->string('nome');
-            $table->integer('desafio');
+            $table->float('desafio');
             $table->foreignId('stats_id')->constrained('stats');
             $table->foreignId('tamanho_id')->constrained('tamanhos');
             $table->foreignId('tipo_id')->constrained('tipos');
